test(types): refactor dtslint tests
