+++ test/types/error.tsx
@@ -1,0 +1,4 @@
+import parse from 'html-react-parser';
+
+// $ExpectError
+parse();
